pyproject.toml
[tool.poetry]
name = "bizztune"
version = "0.1.0"
description = ""
authors = ["ChrisTho23 <christophe.thomassin23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
langfuse = "^2.33.1"
openai = "^1.30.5"
python-dotenv = "^1.0.1"
transformers = "^4.41.2"
mistralai = "^0.3.0"
pandas = "^2.2.2"
sentencepiece = "^0.2.0"
protobuf = "^5.27.1"
torch = "2.2.0"
scikit-learn = "^1.5.0"
datasets = "^2.19.2"
trl = "^0.9.4"
accelerate = "^0.31.0"
peft = "^0.11.1"
wandb = "^0.17.1"

bitsandbytes = "^0.43.1"
evaluate = "^0.4.2"
[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
create_dataset = "bizztune.dataset:main"
benchmark = "bizztune.benchmark:main"
create_instruction = "bizztune.instruction.instruction:main"
